Create platform metadata crate
