+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "mullvad-problem-report",
     "mullvad-jni",
     "mullvad-paths",
+    "mullvad-platform-metadata",
     "mullvad-types",
     "mullvad-rpc",
 #    "mullvad-tests",
