+++ mullvad-platform-metadata/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "mullvad-platform-metadata"
+version = "0.1.0"
+authors = ["Mullvad VPN"]
+description = "Platform metadata detection functions"
+license = "GPL-3.0"
+edition = "2018"
+publish = false
+
+
+[target.'cfg(target_os = "linux")'.dependencies]
+rs-release = "0.1.7"
