+++ mullvad-problem-report/Cargo.toml
@@ -18,16 +18,13 @@ uuid = { version = "0.8", features = ["v4"] }
 tokio = { version = "0.2", features = [ "rt-core" ] }
 
 mullvad-paths = { path = "../mullvad-paths" }
+mullvad-platform-metadata = { path = "../mullvad-platform-metadata" }
 mullvad-rpc = { path = "../mullvad-rpc" }
 talpid-types = { path = "../talpid-types" }
 
 
 [target.'cfg(target_os = "android")'.dependencies]
 duct = "0.13"
-
-
-[target.'cfg(target_os = "linux")'.dependencies]
-rs-release = "0.1.7"
 
 
 [target.'cfg(windows)'.build-dependencies]
